backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.28.1
beautifulsoup4==4.12.3
google-genai==1.10.0
boto3==1.34.84
prometheus-fastapi-instrumentator==6.1.0
